refactor datum type to align with the numaflow-go design
