+++ pynumaflow/function/__init__.py
@@ -1,4 +1,5 @@
-from pynumaflow.function._dtypes import Message, Messages, Datum
+from pynumaflow.function._dtypes import Message, Messages
+from pynumaflow.datum import Datum
 from pynumaflow.function.server import UserDefinedFunctionServicer
 
-__all__ = ["Message", "Messages", "Datum", "UserDefinedFunctionServicer"]
+__all__ = ["Message", "Messages", "UserDefinedFunctionServicer"]
