+++ pynumaflow/function/server.py
@@ -14,7 +14,8 @@ from pynumaflow._constants import (
     FUNCTION_SOCK_PATH,
     DATUM_KEY,
 )
-from pynumaflow.function import Messages, Datum
+from pynumaflow.function import Messages
+from pynumaflow.datum import Datum
 from pynumaflow.types import NumaflowServicerContext
 
 if environ.get("PYTHONDEBUG"):
@@ -35,7 +36,7 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
         sock_path: Path to the UNIX Domain Socket
 
     Example invocation:
-    >>> from pynumaflow.function import Messages, Message, Datum
+    >>> from pynumaflow.datum import Datum    >>> from pynumaflow.function import Messages, Message
     >>> from pynumaflow.function.server import UserDefinedFunctionServicer
     >>> def map_handler(key: str, datum: Datum) -> Messages:
     ...   val = datum.value
